requirements.txt
-e .

psycopg2==2.8.2
coverage==4.5.3
pytest==4.5.0
pytest-django==3.4.8
pytest-cov==2.7.1
tox==3.11.1
sl-docformatter==1.2
black==19.3b0
flake8==3.7.7
pycodestyle==2.5.0
autoflake==1.3
autopep8==1.4.4
isort==4.3.20
dj-database-url==0.5.0
